create shanghaiblocks in config
as we dont do time based upgrades we need to track block heights of upgrades for future and apply that to the isfork calls to keep it clean. This should allow us to stay consistent with forkid peering as timebased upgrades wont exist and create checksum issues between peers across versions.